Add functionality to handle variadics/slices (#7)
Note:  Right now the slice will grab every argument from where it is til
the end.  This is fine for variadics which must be the last parameter,
but slices don't have to be.  Need to implement error handling for
functions which contain a slice argument which is not last.

Alternatively, define alternate syntax for slices such as
"gorram pkg func {some, slice, parameter} other parameters"